backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
python-dotenv==1.0.1
python-multipart==0.0.9
httpx==0.27.0
twilio>=8.0.0
